:sparkles: Feat: Add fork procedures types
