+++ packages/procedures-types/src/procedure/ForkJsonRpcProcedure.ts
@@ -1,0 +1,8 @@
+import type { ForkJsonRpcRequest, ForkJsonRpcResponse } from '../index.js'
+
+/**
+ * Fork JSON-RPC procedure executes a call against the tevm EVM
+ */
+export type ForkJsonRpcProcedure = (
+	request: ForkJsonRpcRequest,
+) => Promise<ForkJsonRpcResponse>
