+++ packages/procedures-types/src/procedure/index.ts
@@ -8,3 +8,4 @@ export type * from './GetAccountJsonRpcProcedure.js'
 export type * from './LoadStateJsonRpcProcedure.js'
 export type * from './SetAccountJsonRpcProcedure.js'
 export type * from './ScriptJsonRpcProcedure.js'
+export type * from './ForkJsonRpcProcedure.js'
