+++ packages/procedures-types/src/requests/ForkJsonRpcRequest.ts
@@ -1,0 +1,6 @@
+import type { JsonRpcRequest } from '@tevm/jsonrpc'
+
+/**
+ * The JSON-RPC request for the `tevm_fork` method
+ */
+export type ForkJsonRpcRequest = JsonRpcRequest<'tevm_fork', {}>
