+++ packages/procedures-types/src/requests/index.ts
@@ -9,3 +9,4 @@ export * from './LoadStateJsonRpcRequest.js'
 export * from './ScriptJsonRpcRequest.js'
 export * from './SetAccountJsonRpcRequest.js'
 export * from './TevmJsonRpcRequest.js'
+export * from './ForkJsonRpcRequest.js'
