+++ packages/procedures-types/src/responses/index.ts
@@ -8,3 +8,4 @@ export * from './GetAccountJsonRpcResponse.js'
 export * from './LoadStateJsonRpcResponse.js'
 export * from './ScriptJsonRpcResponse.js'
 export * from './SetAccountJsonRpcResponse.js'
+export * from './ForkJsonRpcResponse.js'
