+++ packages/procedures-types/src/tevm-request-handler/DebugReturnType.ts
@@ -1,0 +1,13 @@
+import type {
+	DebugTraceCallJsonRpcResponse,
+	DebugTraceTransactionJsonRpcResponse,
+} from '../index.js'
+
+/**
+ * A mapping of `debug_*` method names to their return type
+ */
+export type DebugReturnType = {
+	debug_traceTransaction: DebugTraceTransactionJsonRpcResponse
+	debug_traceCall: DebugTraceCallJsonRpcResponse
+}
+
